Improvements to conditions :: for my no reproduce it: https://github.com/mattyx14/otxserver/issues/495 - Fixed

+++ path_7_x/sources/luascript.cpp
@@ -6360,7 +6360,7 @@ int32_t LuaInterface::luaSetConditionParam(lua_State* L)
 	ConditionParam_t key = (ConditionParam_t)popNumber(L);
 	if(Condition* condition = env->getConditionObject(popNumber(L), loaded))
 	{
-		condition->setParam(key, value);
+		condition->setParam(key, (int32_t)value);
 		lua_pushboolean(L, true);
 	}
 	else
